Removes model sources; removes tests (either outdated or moved to model repository); cleans up packages structure
