+++ application/src/main/scala/cloud/benchflow/experiment/heuristics/jvm/Simple.scala
@@ -1,8 +1,8 @@
 package cloud.benchflow.experiment.heuristics.jvm
 
-import cloud.benchflow.experiment.config.experimentdescriptor.BenchFlowExperiment
 import cloud.benchflow.experiment.heuristics.HeuristicConfiguration
 import cloud.benchflow.driversmaker.utils.env.BenchFlowEnv
+import cloud.benchflow.test.config.experiment.BenchFlowExperiment
 
 /**
   * @author Simone D'Avico (simonedavico@gmail.com)
