+++ application/src/main/scala/cloud/benchflow/experiment/heuristics/scale/ExtendedScaleBalancer.scala
@@ -1,6 +1,7 @@
 package cloud.benchflow.experiment.heuristics.scale
 
-import cloud.benchflow.experiment.config.experimentdescriptor.Driver
+import cloud.benchflow.test.config.experiment.Driver
+
 
 /**
   * @author Simone D'Avico (simonedavico@gmail.com)
