+++ application/src/main/scala/cloud/benchflow/experiment/sources/processors/FabanBenchmarkDeploymentDescriptorProcessor.scala
@@ -2,8 +2,8 @@ package cloud.benchflow.experiment.sources.processors
 
 import java.nio.file.Path
 
-import cloud.benchflow.experiment.config.experimentdescriptor.BenchFlowExperiment
 import cloud.benchflow.driversmaker.utils.env.DriversMakerEnv
+import cloud.benchflow.test.config.experiment.BenchFlowExperiment
 import com.sun.faban.harness.DefaultFabanBenchmark2
 import org.apache.commons.io.{Charsets, FileUtils}
 import spoon.reflect.declaration.CtClass
