+++ application/src/main/scala/cloud/benchflow/experiment/sources/processors/drivers/annotations/BenchmarkDefinitionAnnotation.scala
@@ -1,8 +1,8 @@
 package cloud.benchflow.experiment.sources.processors.drivers.annotations
 
-import cloud.benchflow.experiment.config.experimentdescriptor.BenchFlowExperiment
 import cloud.benchflow.experiment.sources.processors.BenchmarkSourcesProcessor
 import cloud.benchflow.driversmaker.utils.env.DriversMakerEnv
+import cloud.benchflow.test.config.experiment.BenchFlowExperiment
 import com.sun.faban.driver.BenchmarkDefinition
 import spoon.reflect.declaration.CtClass
