--- application/src/test/resources/anotherTest.yaml
@@ -1,28 +1,0 @@
-wfmsbenchmark:
-  description: A WfMS benchmark
-  hostConfig:
-    host: 1.0.0.0
-  scale: 10
-  runControl:
-    rampUp: 30
-    steadyState: 120
-    rampDown: 30
-  drivers:
-    - cloud.benchflow.drivers.WfMSDriver:
-        properties:
-          - property:
-              name: contextPath
-              value: /
-          - property:
-              name: secure
-              value: false
-          - property:
-              name: a
-              value: b
-  sut:
-    - sutconfig:
-        hostConfig:
-        hostPorts: 1.0.0.0:6060
-
-
-
