--- application/src/test/resources/camundaTest.yml
@@ -1,38 +1,0 @@
-benchFlowBenchmark:
-  description: This is the benchmark description
-  hostConfig:
-    host: NNN.NNN.NNN.NN NNN.NNN.NNN.NN NNN.NNN.NNN.NN
-    tools: NONE
-  scale: 100
-  timeSync: false
-  runControl:
-    rampUp: 30
-    steadyState: 100
-    rampDown: 30
-  threadStart:
-    delay: 1000
-    simultaneous: false
-    parallel: false
-  stats:
-    maxRunTime: 6
-    interval: 30
-  drivers:
-    - cloud.benchflow.wfmsbenchmark.driver.WfMSBenchmarkDriver:
-        agents: NNN.NNN.NNN.NN:5 NNN.NNN.NNN.NN:10 NNN.NNN.NNN.NN:10
-        stats:
-          interval: 30
-
-  sut:
-    - sutConfig:
-        hostPorts: NNN.NNN.NNN.NN:MMMM
-        protocol: http
-        contextPath: engine-rest
-        name: camunda
-    - services:
-        benchFlowCompose: http://NNN.NNN.NNN.NN:MMMM
-        collectors:
-          mysqldumpData: http://NNN.NNN.NNN.NN:MMMM/data
-          statsStart: http://NNN.NNN.NNN.NN:MMMM/start
-          statsStop: http://NNN.NNN.NNN.NN:MMMM/stop
-    - runInfo:
-        runID: sut_camunda_1M
