+++ application/src/test/scala/cloud/benchflow/experiment/sources/GenerationTest.scala
@@ -2,9 +2,13 @@ package cloud.benchflow.experiment.sources
 
 import java.nio.file.Paths
 
-import cloud.benchflow.experiment.config.experimentdescriptor.BenchFlowExperiment$
+import cloud.benchflow.driversmaker.requests.Trial
+import cloud.benchflow.experiment.config.FabanBenchmarkConfigurationBuilder
+import cloud.benchflow.experiment.config.deploymentdescriptor.{SiblingVariableResolver, DeploymentDescriptorBuilder}
 import cloud.benchflow.experiment.sources.generators.BenchmarkSourcesGenerator
 import cloud.benchflow.driversmaker.utils.env.{DriversMakerEnv, BenchFlowEnv}
+import cloud.benchflow.test.config.experiment.BenchFlowExperiment
+import cloud.benchflow.test.deployment.docker.compose.DockerCompose
 
 /**
   * @author Simone D'Avico (simonedavico@gmail.com)
@@ -13,56 +17,46 @@ import cloud.benchflow.driversmaker.utils.env.{DriversMakerEnv, BenchFlowEnv}
   */
 object GenerationTest extends App {
 
-  val bb =
-    """
-      |sut:
-      |    name: camunda
-      |    version: v1
-      |    type: wfms
-      |description: A generated benchmark
-      |benchmark_name: MyBenchmark
-      |trials: 2
-      |virtualUsers: 1000
-      |execution:
-      |    rampUp: 0
-      |    steadyState: 60
-      |    rampDown: 0
-      |properties:
-      |    timeSync: "false"
-      |drivers:
-      |    - start:
-      |          operations:
-      |              - myModel:
-      |                    data: ciaociao
-      |          configuration:
-      |              deviation: 5
-      |sut-configuration:
-      |    target-service:
-      |        name: camunda
-      |        endpoint: /engine-rest
-      |    deploy:
-      |        camunda: bull
-      |    benchflow-config:
-      |        camunda: [ stats ]
-    """.stripMargin
 
-  val parsedConfiguration = BenchFlowExperiment.fromYaml(bb)
-
+  val trial = new Trial
+  trial.setBenchmarkId("fooBenchmark")
+  trial.setExperimentNumber(1)
+  trial.setTrialNumber(1)
+  trial.setTotalTrials(3)
   val configYml = new BenchFlowEnv("./application/src/test/resources/app/config.yml")
-//  val benchFlowEnv = new DriversMakerEnv(configYml,
-//    "./application/src/test/resources/app/benchflow-services",
-//    "./application/src/test/resources/app/drivers/templates/skeleton/benchmark")
 
   val benchFlowEnv = new DriversMakerEnv(configYml,
     "./application/src/test/resources/app/benchflow-services",
-    "./application/src/test/resources/app/drivers", 8080)
+    "./application/src/test/resources/app/drivers",
+    8080)
 
-  BenchmarkSourcesGenerator(
-    experimentId = "MyExperiment.1.1",
-    benchFlowBenchmark = parsedConfiguration,
+  val dc = scala.io.Source.fromFile("./application/src/test/resources/docker-compose.yml").mkString
+  val bb = scala.io.Source.fromFile("./application/src/test/resources/benchflow-benchmark.yml").mkString
+  val parsedBB = BenchFlowExperiment.fromYaml(bb)
+
+  val parsedDc = DockerCompose.fromYaml(dc)
+  val parsedTestConfig = BenchFlowExperiment.fromYaml(bb)
+
+  val dcBuilder = new DeploymentDescriptorBuilder(
+    testConfig = parsedTestConfig,
+    env = benchFlowEnv
+  )
+  val resolvedDC = dcBuilder.resolveDeploymentDescriptor(parsedDc, trial)
+  //  println(resolvedDC)
+  //  println(resolvedDC.toYaml.prettyPrint)
+
+  val runXmlBuilder = new FabanBenchmarkConfigurationBuilder(parsedBB,benchFlowEnv,parsedDc)
+  //println(new PrettyPrinter(400, 2).format(runXmlBuilder.build(trial)))
+
+  val siblingResolver = new SiblingVariableResolver(parsedDc, benchFlowEnv, parsedBB)
+
+  val benchmarkSourcesGenerator = BenchmarkSourcesGenerator(
+    experimentId = trial.getExperimentId,
+    benchFlowBenchmark = parsedBB,
     generatedBenchmarkOutputDir = Paths.get("./application/src/test/resources/generated"),
     env = benchFlowEnv
-    //generationResources = Paths.get("./application/src/test/resources/app/drivers")
-  ).generate()
+  )
+
+  benchmarkSourcesGenerator.generate()
 
 }
